+++ test-app/db/migrations/000000003-create-composition-assets.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely'
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('composition_assets')
-    .addColumn('id', 'serial', col => col.primaryKey())
-    .addColumn('composition_id', 'integer', col => col.notNull())
+    .addColumn('id', 'bigserial', col => col.primaryKey())
+    .addColumn('composition_id', 'bigint', col => col.notNull())
     .addColumn('src', 'text')
     .addColumn('primary', 'boolean', col => col.defaultTo(false))
     .addColumn('created_at', 'timestamp', col => col.notNull())
