+++ test-app/db/migrations/000000004-create-composition-asset-audits.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely'
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('composition_asset_audits')
-    .addColumn('id', 'serial', col => col.primaryKey())
-    .addColumn('composition_asset_id', 'integer', col => col.notNull())
+    .addColumn('id', 'bigserial', col => col.primaryKey())
+    .addColumn('composition_asset_id', 'bigint', col => col.notNull())
     .addColumn('approval', 'boolean')
     .addColumn('created_at', 'timestamp', col => col.notNull())
     .addColumn('updated_at', 'timestamp', col => col.notNull())
