+++ test-app/db/migrations/000000006-create-ratings.ts
@@ -3,10 +3,10 @@ import { Kysely, sql } from 'kysely'
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('ratings')
-    .addColumn('id', 'serial', col => col.primaryKey())
-    .addColumn('user_id', 'integer', col => col.references('users.id').onDelete('cascade').notNull())
+    .addColumn('id', 'bigserial', col => col.primaryKey())
+    .addColumn('user_id', 'bigint', col => col.references('users.id').onDelete('cascade').notNull())
     // .addColumn('rateable', 'references', { polymorphic: true }, col => col.references('users.id').onDelete('cascade').notNull())
-    .addColumn('rateable_id', 'integer', col => col.notNull())
+    .addColumn('rateable_id', 'bigint', col => col.notNull())
     .addColumn('rateable_type', 'varchar', col => col.notNull())
     .addColumn('rating', 'integer')
     .addColumn('created_at', 'timestamp', col => col.notNull())
